Added minimal ELF header and non active code to parse symbols.
Added a minimal ELF header which currently only contains specification
for the memory layout of a symbol, and code to parse a symbol table.
Unfortunatelly, the current toolchain (ld, objcopy) does not allow
marking the symbol table as loadable (so that it is loaded by GRUB),
preventing the code from actually being useful until this is worked
around.